Fixed warnings

+++ src/gameObjects/Asteroid.cpp
@@ -116,10 +116,10 @@ namespace GameObjects
 			Rectangle dest{ asteroid.circle.position.x  ,asteroid.circle.position.y,(float)asteroid.texture.width * asteroid.scale / 2,(float)asteroid.texture.height * asteroid.scale / 2 };
 #if _DEBUG
 
-			DrawCircle(asteroid.circle.position.x, asteroid.circle.position.y, asteroid.circle.radius, WHITE);
+			DrawCircle(static_cast<int>(asteroid.circle.position.x), static_cast<int>(asteroid.circle.position.y), asteroid.circle.radius, WHITE);
 #endif
 
-			drawTexture(asteroid.texture, source, dest, { asteroid.texture.width / 2.0f,asteroid.texture.height / 2.0f }, asteroid.rotation, asteroid.scale / 2, WHITE);
+			drawTexture(asteroid.texture, source, dest, { static_cast<float>(asteroid.texture.width) / 2.0f,static_cast<float>(asteroid.texture.height) / 2.0f }, asteroid.rotation, asteroid.scale / 2, WHITE);
 		}
 	}
